apply performance improvements to zabbix server
our present server is running minimal DO specs and runs out of the 512M
ram easily what with so many pollers and mysql and server and agent and
apache.
out of those mysql is by far the biggest and needs to be reigned in to
fit the available memory. also we don't terribly depend on mysql
performance anyway as nothing is profoundly time critical. an OOM killed
server OTOH is an actual problem!