switch to chef zero, it's much cooler than solo. I don't know why though
in all truth though, zero essentially runs a local server whereas solo is
basically a standalone client version. former having the advantage of
being exactly like a chef-server (albeit a local one) whereas latter is
different in that it has no server at all. so, assuming we'll eventually
end up having a server moving closer to zero is the way to go it seems.