divide the partitioning and mount enablement apart form mounting
mounting to be idempotent looks at the mounted devices, little does it know
that those are physical devices but we may be telling it to mount a
symlinked device. so it tries to mount by-id/yolo while sdb is already
mounted and the mount fails. to bypass we explicitly resolve our
symlink to the physical file before trying to mount it