try to provision cloud slaves without ruby
to simplify things it was decided that we'll move away from the custom
build tech and instead rely on distro rubies. obvious consequence is
that we need to keep our servers better upgraded when new versions come
out so as to not fall too far behind and have gems pull support out from
under us. this does however reduce complexity and therefore problems from
busfactor1